requirements.txt
torch>=2.2,<3.0
torchvision>=0.17,<1.0
streamlit>=1.32
pandas>=2.2
numpy>=1.26
pillow>=10.0
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
pyyaml>=6.0
tqdm>=4.66
huggingface_hub>=0.23